refactor: move resubmitBranchesWithNewBases to new file
