+++ src/actions/sync/sync.ts
@@ -1,4 +1,3 @@
-import prompts from 'prompts';
 import { TRepoSyncStackFrame } from '../../lib/config/merge_conflict_callstack_config';
 import { TContext } from '../../lib/context/context';
 import { PreconditionsFailedError } from '../../lib/errors';
@@ -7,16 +6,13 @@ import { syncPRInfoForBranches } from '../../lib/sync/pr_info';
 import {
   checkoutBranch,
   getTrunk,
-  logInfo,
-  logNewline,
-  logTip,
   trackedUncommittedChanges,
 } from '../../lib/utils';
 import { Branch } from '../../wrapper-classes/branch';
 import { deleteMergedBranches } from '../clean_branches';
 import { fixDanglingBranches } from '../fix_dangling_branches';
-import { submitAction } from '../submit';
 import { pull } from './pull';
+import { resubmitBranchesWithNewBases } from './resubmit_branches_with_new_bases';
 
 export async function syncAction(
   opts: {
@@ -86,9 +82,10 @@ export async function repoSyncDeleteMergedBranchesContinuation(
     await resubmitBranchesWithNewBases(frame.force, context);
   }
 
-  const trunk = getTrunk(context).name;
   checkoutBranch(
-    Branch.exists(frame.oldBranchName) ? frame.oldBranchName : trunk
+    Branch.exists(frame.oldBranchName)
+      ? frame.oldBranchName
+      : getTrunk(context).name
   );
 }
 
@@ -108,76 +105,5 @@ function cleanDanglingMetadata(): void {
       ref.delete();
     }
   });
-}*/
-
-async function resubmitBranchesWithNewBases(
-  force: boolean,
-  context: TContext
-): Promise<void> {
-  const needsResubmission: Branch[] = [];
-  Branch.allBranchesWithFilter(
-    {
-      filter: (b) => {
-        const prState = b.getPRInfo()?.state;
-        return (
-          !b.isTrunk(context) &&
-          b.getParentFromMeta(context) !== undefined &&
-          prState !== 'MERGED' &&
-          prState !== 'CLOSED'
-        );
-      },
-    },
-    context
-  ).forEach((b) => {
-    const currentBase = b.getParentFromMeta(context)?.name;
-    const githubBase = b.getPRInfo()?.base;
-
-    if (githubBase && githubBase !== currentBase) {
-      needsResubmission.push(b);
-    }
-  });
-
-  if (needsResubmission.length === 0) {
-    return;
-  }
-
-  logNewline();
-  logInfo(
-    [
-      `The following branches appear to have been rebased (or cherry-picked) in your local repo but changes have not yet propagated to PR (remote):`,
-      ...needsResubmission.map((b) => `- ${b.name}`),
-    ].join('\n')
-  );
-
-  logTip(
-    `Disable this check at any point in the future with --no-resubmit`,
-    context
-  );
-
-  // Prompt for resubmission.
-  let resubmit: boolean = force;
-  if (!force) {
-    const response = await prompts({
-      type: 'confirm',
-      name: 'value',
-      message: `Update PR to propagate local rebase changes? (PR will be re-submitted)`,
-      initial: true,
-    });
-    resubmit = response.value;
-  }
-  if (resubmit) {
-    logInfo(`Updating PR to propagate local rebase changes...`);
-    await submitAction(
-      {
-        scope: 'FULLSTACK',
-        editPRFieldsInline: false,
-        draftToggle: false,
-        dryRun: false,
-        updateOnly: false,
-        branchesToSubmit: needsResubmission,
-        reviewers: false,
-      },
-      context
-    );
-  }
 }
+*/
